Raise exception in LSU when ROB is drained (#108)
This PR adds a notification from ROB to the entire simulation that the ROB
is about to or is expecting simulation to correctly terminate.  If a unit listens
for that notification, it can use that information to determine if simulation 
ended improperly, meaning, it might still have work to do, but didn't expect
simulation to stop.  An example of how this can happen: a unit is NOT 
scheduling an event to do work (a bug) because of a missed path to "wake"
it up.  

If the sparta::Scheduler ends simulation because it has no events to 
run, the unit will not receive the end-of-simulation notification from the ROB
and can throw an exception in this case.

---------

Signed-off-by: Suraj Shirvankar <surajshirvankar@gmail.com>
Co-authored-by: Knute Lingaard <klingaard@gmail.com>