Sparta Assert for Target Pipe without Execution Unit (#151)
Forgot to add an assert to check that there is a dispatcher (thus a
subsequent issue queue/execution unit) that can handle a target pipe of
an instruction. If one forgets to define one, i.e you forget to define
an execution unit to handle `DIV` target pipe, the simulator will end
early, but no error will be thrown.