Initial version of prefetcher API
This API proposes to seperate the prefetcher interface (defined in PrefetcherIF.hpp) with
actual implementation of prefetch engine (defined in PrefetchEngineIF.hpp). This approach
allows the model to rapidly customize the prefetchers with multiple prefetch policies
e.g. We can create stride prefetch engine and use the same engine in L2 Prefetcher as well
as L3 prefetcher.

Prefetchers that depend on external (to prefetcher) state can optionally create an Update Object
deriving it from PrefetcherStateUpdateType and pass the updates to the prefetcher.

NextLinePrefetchEngine is an example implementation of PrefetchEngineIF
DataPrefetcher is an example implementation of a simple prefetcher utilizing the above engine

The files have been compile tested.
Unit testing, LSU integraion is TBD and will be done after review and possible changes.