Added script to auto-generate the rv64v uarch json (#177)
Added a script to auto-generate the rv64v uarch json file using the
Mavis isa json. This will help us keep track of which vector
instructions still need to be implemented in Olympia. Olympia will throw
an exception when trying to construct an `Inst` object for an
unsupported instruction. Added a simple test to make sure that Olympia
fails as expected.