A sample contributor list (#244)
Setting up a contributor list, per request. Note that list is currently
incomplete and is a sample list to solicit feedback. I plan to add every
one who have contributed to this project in that list.

This list is based on the format from this repo
https://github.com/iTowns/itowns/blob/master/CONTRIBUTORS.md?plain=1
(Thanks @jeffnye-gh for suggesting).

Please share your feedback.

Co-authored-by: Knute Lingaard <klingaard@gmail.com>