Flushing SYS instructions on retire should halt same cycle retires (#257)
Failing test and solution for Issue #253 

Some of the SYS instrutions are executed on retire and may cause a
flush. Olympia currently allows the youger uops to retire in the same
cycle.

However, when the flushing sys uop is retired, the program order check
fails, as younger uops have retired out of order.

Solution is to stop retiring uops when this happens.

---------

Co-authored-by: Rajat Bhatia <rbhatia@ventanamicro.com>