trace gen doc update and announce Spike-STF (#258)
For olympia this is a documentation update for trace generation.

Also this PR announces public release of a Spike fork that supports STF
and BBV generation (called Spike-STF). Spike-STF was forked from the
main Spike and the STV/BBV/regression support was developed by me and
others at Condor. There are a number of other features included. BTW:
this version of spike supports the RVA23 profile, which includes vector.

The public repo is here
https://github.com/jeffnye-gh/cpm.riscv-isa-sim.git.

In the PR I have updated the trace/README.md with details on how to
generate traces under linux and baremetal and use them with olympia.

The trace README.md documents how to use the Spike-STF repo, cloning it
within the olympia tree, d/loading a baremetal and linux cross-compiler,
d/loading the kernel, buildroot and opensbi. There is a fair bit of
automation to support linux and build these packages. There are bash
scripts that illustrate how to use the constructed collateral to create
traces for olympia.

I have kept the Dromajo instructions, but I think long term this will
replace Dromajo because Spike's ISA support is current.

I have also added three dhrystone traces generated with Spike-STF. These
three dhrystone align with the accepted practice of 3 levels of
optimization, ground rules, in-lining, and LTO.

---------

Co-authored-by: jeffnye-gh <habanaro@gmail.com>