Docker workflow for STF trace generation (#264)
This PR introduces a new folder, docker_stf_trace_gen, as part of the
ongoing Workload Analysis Flow mentorship. It includes:
* A Dockerfile that builds a container with the RISC-V GNU toolchain,
pk, and Spike-STF.
* Supporting scripts for building and running the Docker image
* Documentation Readme

The current workflow does not yet include the commands to produce STF
files. These will be added in a future PR. For now it only includes the
basis for the documentation and the Dockerfile with the necessary tools.